requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
idna==3.10
pillow==11.1.0
pypdf==5.2.0
requests==2.32.3
urllib3==2.3.0
